+++ R/expr.R
@@ -354,6 +354,9 @@ switch_expr <- function(.x, ...) {
 #' @param x An object or expression to print.
 #' @param width The width of the deparsed or printed expression.
 #'   Defaults to the global option `width`.
+#' @param ... Arguments passed to `expr_deparse()`.
+#' @param max_elements Maximum length of a vector or list before truncation
+#'   occurs. Defaults to 5L.
 #'
 #' @export
 #' @examples
@@ -382,12 +385,19 @@ switch_expr <- function(.x, ...) {
 #'
 #' wrapper_quo <- local(quo(bar(!!local_quo, baz)))
 #' expr_print(wrapper_quo)
-expr_print <- function(x, width = peek_option("width")) {
-  cat_line(expr_deparse(x, width = width))
+expr_print <- function(x, ...) {
+  cat_line(expr_deparse(x, ...))
 }
 #' @rdname expr_print
 #' @export
-expr_deparse <- function(x, width = peek_option("width")) {
-  deparser <- new_quo_deparser(width = width)
+expr_deparse <- function(x,
+                         ...,
+                         width = peek_option("width"),
+                         max_elements = 5L) {
+  check_dots_empty(...)
+  deparser <- new_quo_deparser(
+    width = width,
+    max_elements = max_elements
+  )
   quo_deparse(x, deparser)
 }